wstepne dodanie klas do server
